transactionsx: read a negative ATR expiry instead of failing on it
AtrAttemptJson.ExpiryTimeMillis was unsigned, so a negative "exp" failed to
unmarshal. Nothing stops another client writing one, and the failure was not
confined to that attempt: the lost cleaner unmarshals the whole "attempts" map in
one go, so a single entry with a negative expiry made every attempt in that ATR
unreadable and none of them was ever cleaned up.

A negative value is not even nonsense - it says the attempt was already expired
when it wrote its pending entry, which is exactly the condition cleanup exists
for, and the arithmetic already handles it: an expiry that far in the past is
unconditionally past due, so the entry is picked up on the first pass.

Signed now, for that reason.

Caught by CleanupBadExpirationTimesTest.negativeExpirationTimeOnPendingTransactionShouldBeCleanedUp,
which forces exp to -20000000 and expects prompt cleanup.

+++ transactionsx/xattr_types.go
@@ -47,7 +47,13 @@ type AtrAttemptJson struct {
 	// "exp" and the unit setATRPending writes.  It is measured from PendingCAS,
 	// not from when the entry is read, so an absolute expiry time is
 	// PendingCAS-as-a-time plus this.
-	ExpiryTimeMillis uint `json:"exp,omitempty"`
+	//
+	// Signed, because nothing stops another client writing a negative value and
+	// an unsigned field fails to unmarshal one -- taking the whole ATR with it,
+	// so one bad entry hides every other attempt in that ATR from cleanup.  A
+	// negative value just means the attempt was already expired when it was
+	// written, which is coherent and actionable.
+	ExpiryTimeMillis int64 `json:"exp,omitempty"`
 
 	State TxnStateJson `json:"st,omitempty"`
